add new controller and route

+++ controllers/Kategoris.js
@@ -3,9 +3,9 @@ import Kategori from "../models/KategoriModel.js";
 export const getKategoris = async (req, res) => {
     try {
         const kategoris = await Kategori.findAll();
-        res.json(kategoris);
+        res.status(200).json(kategoris);
     } catch (error) {
-        console.log(error);
+        res.status(500).json({msg: error.message});
     }
 }
 
@@ -13,52 +13,65 @@ export const getKategoriById = async (req, res) => {
     try {
         const kategori = await Kategori.findOne({
             where: {
-                id: req.params.id
+                uuid: req.params.id
             }
         });
-        res.json(kategori);
+        res.status(200).json(kategori);
     } catch (error) {
-        console.log(error);
+        res.status(500).json({msg: error.message});
     }
 }
 
 export const createKategori = async (req, res) => {
+    const {name} = req.body;
     try {
-        await Kategori.create(req.body);
-        res.json({
-            message: "Kategori created successfully"
-        });
+        await Kategori.create({
+            name: name
+        });   
+        res.status(201).json({msg: "Kategori Created"});     
     } catch (error) {
-        console.log(error);
+        res.status(400).json({msg: error.message});
     }
 }
 
 export const updateKategori = async (req, res) => {
-    try {
-        await Kategori.update(req.body, {
-            where: {
-                id: req.params.id
-            }
-        });
-        res.json({
-            message: "Kategori updated successfully"
-        });
-    } catch (error) {
-        console.log(error);
+   const kategori = await Kategori.findOne({
+       where: {
+           uuid: req.params.id
+       }
+   });
+
+   const {name} = req.body;
+   try {
+    await Kategori.update({
+        name: name
+    }, {
+        where: {
+            id: kategori.id
+        }
+    });
+    res.status(200).json({msg: "Kategori Updated"});
+   } catch (error) {
+    res.status(400).json({msg: error.message});
     }
+
 }
 
 export const deleteKategori = async (req, res) => {
+    const kategori = await Kategori.findOne({
+        where: {
+            uuid: req.params.id
+        }
+    });
+  if(!kategori) return res.status(404).json({msg: "No Data Found"});
     try {
-        await Kategori.destroy({
-            where: {
-                id: req.params.id
-            }
-        });
-        res.json({
-            message: "Kategori deleted successfully"
-        });
+     await Kategori.destroy({
+         where: {
+             id: kategori.id
+         }
+     });
+     res.status(200).json({msg: "Kategori Deleted"});
     } catch (error) {
-        console.log(error);
-    }
+     res.status(400).json({msg: error.message});
+     }
 }
